working on board progress - board logic moved into child component board.js from user-home
